Buggy code and .txt file for documentation
In regards to the buggy code, I tried making an empty map and then trying to write a function that moves the player around. The switch case statement doesn't seem to work, as "this constant expression has type "const char *" instead of the required "int" type".

For documentation, I suggest we all use a .txt file to properly document everyting at the end of the project, so life is more bearable.