Lots of new stuff for rendering goblins from array
So basically, I changed the way the map is rendered FROM SCRATCH, also removed an unnecessary line from Character.h and the preplaced goblin from Character.cpp. Combat is now fully functional, complete with not moving if the space that the player is moving into is occupied by a goblin.